requirements.txt
absl-py==1.3.0
adversarial-robustness-toolbox==1.12.1
astroid==2.11.7
astunparse==1.6.3
autoattack==0.1
black==22.6.0
botorch==0.6.2
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.0
compressai==1.2.1
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.30
cytoolz==0.11.0
dask==2022.10.1
dill==0.3.4
docker-pycreds==0.4.0
eagerpy==0.30.0
efficientnet-pytorch==0.7.1
flake8==4.0.1
flatbuffers==22.10.26
fonttools==4.25.0
foolbox==3.3.3
fsspec==2022.10.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpytorch==1.9.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imagecodecs==2021.8.26
imagecorruptions==1.1.2
imageio==2.22.0
importlib-metadata==4.11.3
isort==5.9.3
joblib==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kornia==0.6.8
lazy-object-proxy==1.6.0
libclang==14.0.6
linear-operator==0.0.0
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mccabe==0.7.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munch==2.5.0
munkres==1.1.4
mypy==0.910
mypy-extensions==0.4.3
networkx==2.8.7
numba==0.56.3
numexpr==2.8.3
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
partd==1.3.0
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.2.0
pip==22.2.2
platformdirs==2.5.2
ply==3.11
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.14.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.9.1
segmentation-models-pytorch==0.3.0
sentry-sdk==1.10.1
setproctitle==1.2.2
setuptools==65.5.0
shortuuid==1.0.9
sip==6.6.2
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==2.2.0
tifffile==2021.4.8
timm==0.4.12
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
torch==1.13.0
torchaudio==0.13.0
torchjpeg==0.9.27
torchmetrics==0.10.2
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.5
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.0
